pyproject.toml
[tool.poetry]
name = "projects-manager"
version = "0.1.0"
description = "Simple backend of project manager platform"
authors = ["maciej-MKan"]
license = "None"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyramid = "2.0"
pyramid-tm = "^2.5"
pyramid-retry = "^2.1.1"
psycopg2-binary = "^2.9.6"
pyramid-sqlalchemy = "^1.6"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
dependency-injector = "^4.41.0"

[tool.poetry.group.dev.dependencies]
WebTest = "^3.0.0"
mypy = "^1.2.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
